requirements.txt
ask-sdk==1.17.1
python-miio==0.5.12
micloud==0.5